refactoring: revisited maven samples
