+++ maven/dagger-example/src/main/kotlin/examples/dagger/CoffeeMaker.kt
@@ -1,0 +1,18 @@
+package examples.dagger
+
+import dagger.Lazy
+import javax.inject.Inject
+
+class CoffeeMaker @Inject constructor(
+        private val heater: Lazy<Heater>,
+        private val pump: Pump
+) {
+
+    fun brew() {
+        heater.get().on()
+        pump.pump()
+        println(" [_]P coffee! [_]P ")
+        heater.get().off()
+    }
+
+}
