+++ maven/dagger-example/src/main/kotlin/examples/dagger/DripCoffeeModule.kt
@@ -1,0 +1,16 @@
+package examples.dagger
+
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
+
+@Module(includes = [PumpModule::class])
+class DripCoffeeModule {
+
+    @Provides
+    @Singleton
+    fun provideHeater(): Heater {
+        return ElectricHeater()
+    }
+
+}
