+++ maven/dagger-example/src/main/kotlin/examples/dagger/ElectricHeater.kt
@@ -1,0 +1,16 @@
+package examples.dagger
+
+open class ElectricHeater : Heater {
+
+    override var isHot: Boolean = false
+
+    override fun on() {
+        println("~ ~ ~ heating ~ ~ ~")
+        this.isHot = true
+    }
+
+    override fun off() {
+        this.isHot = false
+    }
+
+}
