+++ maven/dagger-example/src/main/kotlin/examples/dagger/Heater.kt
@@ -1,0 +1,9 @@
+package examples.dagger
+
+interface Heater {
+
+    fun on()
+    fun off()
+    val isHot: Boolean
+
+}
