+++ maven/dagger-example/src/main/kotlin/examples/dagger/Pump.kt
@@ -1,0 +1,7 @@
+package examples.dagger
+
+interface Pump {
+
+    fun pump()
+
+}
