+++ maven/dagger-example/src/main/kotlin/examples/dagger/PumpModule.kt
@@ -1,0 +1,12 @@
+package examples.dagger
+
+import dagger.Binds
+import dagger.Module
+
+@Module
+abstract class PumpModule {
+
+    @Binds
+    abstract fun providePump(pump: Thermosiphon): Pump
+
+}
