+++ maven/dagger-example/src/main/kotlin/examples/dagger/Thermosiphon.kt
@@ -1,0 +1,15 @@
+package examples.dagger
+
+import javax.inject.Inject
+
+class Thermosiphon @Inject constructor(
+        private val heater: Heater
+) : Pump {
+
+    override fun pump() {
+        if (heater.isHot) {
+            println("=> => pumping => =>")
+        }
+    }
+
+}
