--- maven/dagger-maven-example/.gitignore
@@ -1,6 +1,0 @@
-target
-.idea
-
-.settings/
-.classpath
-.project
