--- maven/dagger-maven-example/src/main/kotlin/ElectricHeater.kt
@@ -1,14 +1,0 @@
-package coffee
-
-open class ElectricHeater : Heater {
-    override var isHot: Boolean = false
-
-    override fun on() {
-        println("~ ~ ~ heating ~ ~ ~")
-        this.isHot = true
-    }
-
-    override fun off() {
-        this.isHot = false
-    }
-}
