--- maven/dagger-maven-example/src/main/kotlin/Heater.kt
@@ -1,7 +1,0 @@
-package coffee
-
-interface Heater {
-    fun on()
-    fun off()
-    val isHot: Boolean
-}
