--- maven/dagger-maven-example/src/main/kotlin/Pump.kt
@@ -1,5 +1,0 @@
-package coffee
-
-interface Pump {
-    fun pump()
-}
