--- maven/dagger-maven-example/src/main/kotlin/PumpModule.kt
@@ -1,10 +1,0 @@
-package coffee
-
-import dagger.Binds
-import dagger.Module
-
-@Module
-abstract class PumpModule {
-    @Binds
-    abstract fun providePump(pump: Thermosiphon): Pump
-}
