--- maven/dokka-maven-example/Module.md
@@ -1,7 +1,0 @@
-# Module dokka-maven-example
-
-This is an example of how you can write module documentation with Dokka.
-
-# Package demo
-
-This package contains a few examples of Dokka usage.
