--- maven/hello-world/.gitignore
@@ -1,2 +1,0 @@
-target
-.idea
