--- maven/hello-world/src/main/kotlin/Hello.kt
@@ -1,10 +1,0 @@
-package hello
-
-fun getHelloString() : String {
-    return "Hello, world!"
-}
-
-fun main(args : Array<String>) {
-    println(getHelloString())
-}
-
