+++ maven/hello-world/src/main/kotlin/examples/helloworld/Hello.kt
@@ -1,0 +1,9 @@
+package examples.helloworld
+
+fun getHelloString() : String {
+    return "Hello, world!"
+}
+
+fun main() {
+    println(getHelloString())
+}
