--- maven/hello-world/src/test/kotlin/HelloTest.kt
@@ -1,11 +1,0 @@
-package hello.tests
-
-import hello.getHelloString
-import kotlin.test.assertEquals
-import org.junit.Test
-
-class HelloTest {
-    @Test fun testAssert() : Unit {
-        assertEquals("Hello, world!", getHelloString())
-    }
-}
