+++ maven/hello-world/src/test/kotlin/examples/helloworld/HelloTest.kt
@@ -1,0 +1,13 @@
+package examples.helloworld
+
+import org.junit.Assert
+import org.junit.Test
+
+class HelloTest {
+
+    @Test
+    fun shouldReturnCorrectString() {
+        Assert.assertEquals("Hello, world!", getHelloString())
+    }
+
+}
