+++ maven/interop-example/src/main/java/examples/interop/JavaHello.java
@@ -1,0 +1,11 @@
+package examples.interop;
+
+public class JavaHello {
+
+    public static String JavaHelloString = "Hello from Java!";
+
+    public static String getHelloStringFromKotlin() {
+        return KotlinHelloKt.getKotlinHelloString();
+    }
+
+}
