+++ maven/interop-example/src/main/kotlin/examples/interop/KotlinHello.kt
@@ -1,0 +1,7 @@
+package examples.interop
+
+val KotlinHelloString : String = "Hello from Kotlin!"
+
+fun getHelloStringFromJava() : String {
+    return JavaHello.JavaHelloString!!;
+}
