--- maven/kotlin-querydsl/.gitignore
@@ -1,11 +1,0 @@
-.DS_Store
-
-generated
-generated_tests
-
-.idea/
-*.iml
-*.iws
-*.ipr
-
-target/
