--- maven/mixed-code-hello-world/.gitignore
@@ -1,6 +1,0 @@
-target
-.idea
-
-.settings/
-.classpath
-.project
