--- maven/mixed-code-hello-world/src/main/kotlin/hello/KotlinHello.kt
@@ -1,7 +1,0 @@
-package hello
-
-val KotlinHelloString : String = "Hello from Kotlin!"
-
-fun getHelloStringFromJava() : String {
-    return JavaHello.JavaHelloString!!;
-}
