+++ maven/querydsl-example/src/main/kotlin/examples/querydsl/User.kt
@@ -1,0 +1,12 @@
+package examples.querydsl
+
+import javax.persistence.Entity
+import javax.persistence.GeneratedValue
+import javax.persistence.GenerationType
+import javax.persistence.Id
+
+@Entity
+class User(
+    val name: String,
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY) val id: Int = 0
+)
